feat: request storage permission in setup
