+++ app/src/main/java/com/osfans/trime/ui/setup/SetupPage.kt
@@ -3,8 +3,12 @@ package com.osfans.trime.ui.setup
 import android.content.Context
 import com.osfans.trime.R
 import com.osfans.trime.util.InputMethodUtils
+import com.osfans.trime.util.appContext
+import com.osfans.trime.util.isStorageAvailable
+import com.osfans.trime.util.requestExternalStoragePermission
 
 enum class SetupPage {
+    Permissions,
     Enable,
     Select,
     ;
@@ -12,14 +16,16 @@ enum class SetupPage {
     fun getStepText(context: Context) =
         context.getText(
             when (this) {
-                Enable -> R.string.setup__step_one
-                Select -> R.string.setup__step_two
+                Permissions -> R.string.setup__step_one
+                Enable -> R.string.setup__step_two
+                Select -> R.string.setup__step_three
             },
         )
 
     fun getHintText(context: Context) =
         context.getText(
             when (this) {
+                Permissions -> R.string.setup__request_permmision_hint
                 Enable -> R.string.setup__enable_ime_hint
                 Select -> R.string.setup__select_ime_hint
             },
@@ -28,6 +34,7 @@ enum class SetupPage {
     fun getButtonText(context: Context) =
         context.getText(
             when (this) {
+                Permissions -> R.string.setup__request_permmision
                 Enable -> R.string.setup__enable_ime
                 Select -> R.string.setup__select_ime
             },
@@ -35,6 +42,7 @@ enum class SetupPage {
 
     fun getButtonAction(context: Context) {
         when (this) {
+            Permissions -> context.requestExternalStoragePermission()
             Enable -> InputMethodUtils.showImeEnablerActivity(context)
             Select -> InputMethodUtils.showImePicker()
         }
@@ -42,17 +50,18 @@ enum class SetupPage {
 
     fun isDone() =
         when (this) {
+            Permissions -> appContext.isStorageAvailable()
             Enable -> InputMethodUtils.checkIsTrimeEnabled()
             Select -> InputMethodUtils.checkIsTrimeSelected()
         }
 
     companion object {
-        fun SetupPage.isLastPage() = this == values().last()
+        fun SetupPage.isLastPage() = this == entries.last()
 
-        fun Int.isLastPage() = this == values().size - 1
+        fun Int.isLastPage() = this == entries.size - 1
 
-        fun hasUndonePage() = values().any { !it.isDone() }
+        fun hasUndonePage() = entries.any { !it.isDone() }
 
-        fun firstUndonePage() = values().firstOrNull { !it.isDone() }
+        fun firstUndonePage() = entries.firstOrNull { !it.isDone() }
     }
 }
