+++ app/src/main/res/values-zh-rCN/strings.xml
@@ -277,4 +277,7 @@
     <string name="pref_keyboard__landscape">横屏模式</string>
     <string name="keyboard__split_title">转换到横屏键盘</string>
     <string name="keyboard__split_space_title">自动分割键盘比例</string>
+    <string name="setup__step_three">第 3 步</string>
+    <string name="setup__request_permmision_hint">当前同文需要存储权限以访问配置、使得用户容易更改。</string>
+    <string name="setup__request_permmision">请求存储权限</string>
 </resources>
