+++ gradle.properties
@@ -4,3 +4,5 @@ android.useAndroidX=true
 
 # enlarge heap size to 2GB
 org.gradle.jvmargs=-Xmx2048M -XX:+UseParallelGC
+
+buildABI=x86,arm64-v8a
